[10/17/2012] modified   SAT/Base.h 1. [ADD] Add the parameter _Reserved to UGetNrOfBins() and UGetNrOfDims().
modified   SAT/CMakeLists.txt
modified   SAT/WaveletSAT.h
1. [MOD] Change the project name to SAT.

modified   SAT/SimpleND.h
1. [ADD] Include IntegralHistogram.h.

modified   SimpleND/SimpleND.cpp
1. [MOD] When assign the value, the data type of the index is changed to int.

+++ SimpleND/SimpleND.cpp
@@ -85,10 +85,14 @@ main(int argn, char* argv[])
 
 	// create a lookup table to shuffule the value
 	vector<int> viShuffleTable;
-	for(size_t i = 0; i < iValueMax; i++)
+	// MOD-BY-LEETEN 10/17/2012-FROM:	for(size_t i = 0; i < iValueMax; i++)
+	for(int i = 0; i < iValueMax; i++)
+	// MOD-BY-LEETEN 10/17/2012-END
 		viShuffleTable.push_back((int)i);
 
-	for(size_t i = iValueMax; i > 0; i--)
+	// MOD-BY-LEETEN 10/17/2012-FROM:	for(size_t i = iValueMax; i > 0; i--)
+	for(int i = iValueMax; i > 0; i--)
+	// MOD-BY-LEETEN 10/17/2012-END
 	{
 		size_t uRand = rand() % i;
 		swap(viShuffleTable[i - 1], viShuffleTable[uRand]);
